Splitted Ubuntu/GNUstep/Etoile generic build stuff out of livecd-build.sh into ubuntu-install-etoile.sh. Improved gdm.conf related handling in case the script is run another time. Improved ld.so.conf update with identical expectation in mind
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@2120 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ LiveCD/Subscripts/livecd-build.sh
@@ -1,76 +1,7 @@
 #!/bin/sh
 
-# --build
-
-mkdir build
-cd build
-
-# Install Etoile and GNUstep dependencies
-
-sudo apt-get -y install gobjc-4.1 openssl libssl-dev libxml2-dev libxslt1-dev libffi4-dev libjpeg62-dev libtiff4-dev libpng12-dev libungif4-dev libfreetype6-dev libx11-dev libart-2.0-dev libxft-dev libxmu-dev libxss-dev xscreensaver libdbus-1-dev libstartup-notification0-dev g++ libpoppler-dev
-
-# Install Subversion to be able to check both GNUstep and Etoile stable versions
-
-sudo apt-get -y install subversion
-
-# Check out and build GNUstep stable version
-
-svn co http://svn.gna.org/svn/gnustep/tools/make/branches/stable gnustep-make
-svn co http://svn.gna.org/svn/gnustep/libs/base/branches/stable gnustep-base
-svn co http://svn.gna.org/svn/gnustep/libs/gui/branches/stable gnustep-gui
-svn co http://svn.gna.org/svn/gnustep/libs/back/branches/stable gnustep-back
-
-cd gnustep-make
-./configure --prefix=/ && make && sudo make install
-cd ..
-. /System/Library/Makefiles/GNUstep.sh
-cd gnustep-base
-./configure && make && sudo make install
-cd ..
-cd gnustep-gui
-./configure && make && sudo make install
-cd ..
-cd gnustep-back
-./configure && make && sudo make install
-cd ..
-
-# Download latest StepTalk release over FTP, uncompress and build it
-
-wget ftp://ftp.gnustep.org/pub/gnustep/libs/StepTalk-0.10.0.tar.gz
-tar -xzf StepTalk-0.10.0.tar.gz
-cd StepTalk
-make && sudo make install
-cd ..
-
-# Download latest Gorm release over FTP, uncompress and build it
-
-wget ftp://ftp.gnustep.org/pub/gnustep/dev-apps/gorm-1.2.1.tar.gz
-tar -xzf gorm-1.2.1.tar.gz
-
-cd gorm-1.2.1
-make && sudo make install
-cd ..
-
-# Check out and build Etoile stable version
-
-svn co http://svn.gna.org/svn/etoile/trunk/Dependencies Dependencies
-svn co http://svn.gna.org/svn/etoile/stable/Etoile Etoile
-
-cd Dependencies/oniguruma5
-./configure && make && sudo make install # Build and install Oniguruma first
-cd ../..
-cd Etoile
-make && sudo make install
-
-# Register /usr/local/lib so libonig can be found by OgreKit
-echo "/usr/local/lib" >> /etc/ld.so.conf
-ldconfig
-
-# Workaround install bug (probably related to gnustep-make)
-# FIXME: Remove this hack.
-
-ln -s /Library/StepTalk /Local/Library/StepTalk
-#ln -s /Library/Grr /Local/Library/Grr
+# Download and install Etoile, GNUstep and all required dependencies
+$SUBSCRIPT_DIR/ubuntu-install-etoile.sh
 
 #
 # In all the code that follows it's important to take in account it can be run
@@ -81,15 +12,17 @@ ln -s /Library/StepTalk /Local/Library/StepTalk
 # Login panel specific stuff
 # NOTE: Keep a copy of original gdm.conf before deleting it to force 
 # gdm.conf-custom to be used
-cp /etc/gdm/gdm.conf /etc/gdm.conf-original 
-rm /etc/gdm/gdm.conf
+if [ -f /etc/gdm.conf ]; then
+	cp /etc/gdm/gdm.conf /etc/gdm.conf-original
+	rm /etc/gdm/gdm.conf
+fi
 sed -e '/^Greeter.*$/d' /etc/gdm/gdm.conf-custom
 sed -e 's/\(^\[greeter\].*$\)/\1\nGreeter=\/usr\/local\/bin\/etoile_login.sh/' /etc/gdm/gdm.conf-custom
 # NOTE: su gdm doesn't work, so we pass the default in GDM greeter script
 #su gdm
 #defaults write Login ETAllowUserToChooseEnvironment 'NO'
 cp Services/Private/Login/etoile_login.sh /usr/local/bin
-sed -e '/-ETAllowUserToChooseEnvironment "NO"/d' /usr/local/bin/etoile_login.sh
+#sed -e '/-ETAllowUserToChooseEnvironment "NO"/d' /usr/local/bin/etoile_login.sh
 sed -e 's/\(Login.*\)/\1 -ETAllowUserToChooseEnvironment "NO"/' /usr/local/bin/etoile_login.sh
 exit
 
@@ -102,9 +35,9 @@ adduser $ETOILE_USER_NAME audio floppy cdrom dialout adm
 # Add 'etoile' user to sudoers
 adduser $ETOILE_USER_NAME admin
 if [ ! -f /etc/sudoers-original ]; then
-	cp /etc/sudoers /etc/sudoers-original;
+	cp /etc/sudoers /etc/sudoers-original
 else
-	cp /etc/sudoers-original /etc/sudoers;
+	cp /etc/sudoers-original /etc/sudoers
 fi
 echo '%admin	ALL=(ALL) ALL' >> /etc/sudoers
